mgf_spectra_plot/requirements.txt
matplotlib
pandas
requests-cache
spectrum-utils
streamlit